+++ York GC Members App/MHSystems/app/src/main/java/com/mh/systems/york/fragments/MyDetailsFragment.java
@@ -107,6 +107,10 @@ public class MyDetailsFragment extends Fragment {
             ((YourAccountActivity) getActivity()).updateFilterIcon(0);
             ((YourAccountActivity) getActivity()).setiOpenTabPosition(0);
 
+            /* ++++++++++++++++  START OF PURSE API FEATURE ++++++++++++++++ */
+            ((YourAccountActivity) getActivity()).setiBalanceType(0);
+            /* ++++++++++++++++  END OF PURSE API FEATURE ++++++++++++++++ */
+
             /**
              *  Check internet connection before hitting server request.
              */